disable room delete button if room is occupied

+++ client/src/components/RoomModal/index.js
@@ -57,6 +57,8 @@ export default function RoomModal(props) {
     rate: "",
     capacity: "",
   });
+  const [rooms, setRooms] = useState([]);
+  const [status, setStatus] = useState(true);
 
   const [markers, setMarkers] = useState({
     number: false,
@@ -136,6 +138,16 @@ export default function RoomModal(props) {
         capacity: props.selected.capacity,
       });
       // setGuest(props.selected.guesyId);
+      API.getRoomInfo().then((res) => {
+        setRooms(res.data.map((room) => room));
+        let room = res.data.filter(({ _id }) => _id === props.selected.id)[0];
+        console.log("DEL:", room);
+        setStatus(room.guests.length !== 0);
+
+        console.log(status);
+      });
+    } else {
+      clearForm();
     }
   }, [props.type, props.selected]);
 
@@ -145,9 +157,15 @@ export default function RoomModal(props) {
   }
 
   function deleteRoom() {
-    API.deleteRoom(props.selected.id);
-    clearForm();
-    props.close();
+    let room = rooms.filter(({ _id }) => _id === props.selected.id)[0];
+    console.log("DEL:", room);
+    if (room.guests.length === 0) {
+      API.deleteRoom(props.selected.id);
+      clearForm();
+      props.close();
+    } else {
+      console.log("Can't delete occupied rooms");
+    }
   }
 
   const firstBody = (
@@ -221,7 +239,9 @@ export default function RoomModal(props) {
       </form>
       <Button onClick={submitForm}>Submit</Button>
       {props.type === "Update" ? (
-        <Button onClick={deleteRoom}>Delete Room</Button>
+        <Button onClick={deleteRoom} disabled={status}>
+          Delete Room
+        </Button>
       ) : (
         <></>
       )}
